requirements.txt
fastapi[standard]==0.115.12
uvicorn==0.34.3
jinja2==3.1.6
aiofiles==24.1.0
httpx==0.28.1
pandas==2.3.0
python-multipart==0.0.20
requests==2.32.4
python-dotenv==1.1.0